Desktop/results-server-main - Copy/results-server-main/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
psycopg2==2.9.4
drf-spectacular==0.24.2
uwsgi>=2.0.19,<2.1
django-rest-durin>=1.0.0,<2.0
django-cors-headers>=3.13.0,<3.14.0
googlemaps>=4.6.0, <4.7.0
Pillow>=9.2.0,<9.3.0
djangorestframework-csv>=2.1.1, <2.2.0
